==> 1.    Cleaning.ipynb <==
import json
import numpy as np
import pandas as pd
from pymongo import MongoClient
import src.filtering as fn
import src.query_fun as mf

client = MongoClient("mongodb://localhost/companies")
db = client.get_database()

### Obtenemos de la colección "companies" las oficinas que cuentan con los apartados:
#### funding_rounds, offices, latitude, longitude, deadpooled_year (que no sea mayor o igual que 0 )

companies = list(db["companies"].find({"funding_rounds":{"$exists": True, "$not": {"$size": 0}},"offices":{"$exists": True, "$not": {"$size": 0}}, "offices.latitude":{"$exists":True, "$ne":None}, "offices.longitude":{"$exists":True, "$ne":None},"deadpooled_year":{"$not":{"$gte":0}}},{"name":1,"offices":1,"category_code":1, "funding_rounds":1, "founded_year":1}))
df = pd.DataFrame(companies)
df.head()

#Encontramos los diferentes códigos de los países y elegimos Canadá
#query={'offices': { '$exists': 'true', '$not': {'$size': 0} }}
#web_offices = list(db.companies.find(query,{"offices":1,"name":1}))

#countries=set()
#for i,e in enumerate(web_offices):
    #countries.add(e['offices'][0]['country_code'])
#countries #CAN es el de Canadá

#### Sumamos el total de funding_rounds y obtenemos la columna total_funding para saber el total de dinero invertido en USD.

total_funding = fn.sum_funding(df.funding_rounds)
df["total_funding"] = total_funding
df.head()

#### Utilizamos "explode" para transformar cada elemento de la lista de la fila offices a distintas filas, replicando los valores índice y, por otro lado utilizamos result_type = "expand" para transformar los resultados "list-like" en columnas.
#### Al concatenar todo obtenemos un dataframe formado por más columnas que contienen toda la información que necesitamos.

df = df.explode('offices')
df_offices= df[["offices"]].apply(lambda r: r.offices, result_type="expand", axis=1)
df = pd.concat([df,df_offices], axis=1)
df.head()

#### Eliminamos las columnas que no nos interesan

df = df.drop(columns = ["funding_rounds", "description", "address2", "state_code", "country_code"])

#### Creamos la columna location como un diccionario de tipo "point" para poder realizar geoqueries con MongoDB.

df["location"] = df[["latitude","longitude"]].apply(lambda x:fn.toGeoJSON(x.latitude,x.longitude), axis=1)
df.head()

#### Vuelvo a eliminar columnas que me doy cuenta de que tampoco son necesarias

df = df.drop(columns=["_id","address1","zip_code"])
df.head()

#### Creamos un json que almacenamos en la carpeta OUTPUT y lo importamos a Mongo DB como una nueva bbdd "Cleaned_offices" con la correspondiente colección llamada  "cleaned_offices".

#df.to_json("cleaned_offices.json", orient="records")

#Incorporamos nueva colección a la bbdd y creamos el índice geoloc 
#!mongoimport --db companies --collection cleaned_offices --drop --jsonArray output/cleaned_offices.json
