==> 2.     Obtaining coordinates and making ranking.ipynb <==
from pymongo import MongoClient
import pandas as pd

import src.googla as gl
import os
from dotenv import load_dotenv
import json
import requests
import src.filtering as fl

load_dotenv()

conn = MongoClient("localhost:27017")

#### Comprobamos que existe y aparece la base de datos "Cleaned_offices" que necesitamos utilizar, así como la colección correspondiente

conn.list_database_names()

db = conn.get_database("Cleaned_offices")

db.list_collection_names()

collection = db.get_collection("cleaned_offices")


collection.find_one()


#### Comprobamos las keys que hay 

collection.find_one().keys()


###  Después de investigar al respecto  e informarme acerca de la situación tecnológica de algunas de las economías más fuertes del mundo, he decidido situar la oficina en Vancouver, Canadá. Esto se debe, por un lado, a que[ el sector tecnológico de la ciudad está en pleno auge](https://www.theglobeandmail.com/business/industry-news/property-report/article-vancouvers-emerging-tech-sector-poised-for-accelerated-growth/), llamando la atención de empresas como Microsoft Corp.,  Sony Pictures Imageworks Inc. o Amazon.  Además, por otra parte, se espera que [la fuerte economía en crecimiento de Vancouver en la situación pre-COVID se recupere casi en su totalidad](https://www.vancouvereconomic.com/economic-snapshot/), cuestión interesante teniendo en cuenta la crisis económica que sucederá a la sanitaria que se está viviendo a nivel mundial.

can_offices = collection.find({"offices.country_code": 'CAN', "city": "Vancouver"})

offices_list = list(can_offices)

offices_list[0]

#def get_google(latlong, radius,typ, apiKey, keyw=None):
  #  keyword= '&keyword='+keyw if keyw else ''
  #  url = f'https://maps.googleapis.com/maps/api/place/nearbysearch/json?location={latlong}&radius={radius}&type={typ}{keyword}&key={apiKey}'
 #   res = requests.get(url)
 #   return res.json()['results']

#def get_latlong(office):
  #  return str(office['location']['coordinates'][::-1]).replace(' ', '')[1:-1]

#### Utilizamos una función ("fill_nearest_places") para llamar a la API de Google Places y, a partir de la lista de oficinas filtradas en el paso anterior,  obtener de cada oficina los aeropuertos, los Starbucks, los clubs nocturnos y los colegios y realizamos un "len" de la lista para así saber cuántos de cada tiene cada oficina en el radio que le marcamos:

-  Aeropuertos (internacionales) en un radio de 2 km. Especificamos internacionales porque Vancouver tiene algunos aeropuertos para sea-planes.y así poder viajar entre islas cercanas.

- Colegios en un radio de 1 km.

- Clubs nocturnos en un radio de 1 km.

- Starbucks en un radio de 500 metros.


filled_offices_list = gl.fill_nearest_places(offices_list)

filled_offices_list[0]

filled_offices_df = pd.DataFrame(filled_offices_list)

filled_offices_df.head(3)

#### Ahora procedemos a ponderar. Para mi análisis he considerado que:
- Tener un aeropuerto internacional en un radio de 2 km tiene un peso de 0,3 porque al ganar tiempo minimiza costes para la empresa.
- Tener un colegio en un radio de 1 km tiene un peso de 0,2, porque el 30% de la plantilla tiene hijos.
- Tener un club nocturno a menos de 1 km tiene un peso de 0,3, porque teniendo en cuenta la edad de la plantilla podría interesar a todos.
- Tener un Starbucks a menos de 500 metros tiene un peso de 0,2 porque es importante que los ejecutivos no pierdan demasiado tiempo de trabajo yendo a por café.

#### Con esta información obtenemos la columna 'score' que ordenamos de forma descendente de modo que los primeros resultados son los de las oficinas que en base a los criterios explicados han obtenido una mayor puntuación.

filled_offices_df['score'] = (filled_offices_df["airports_2km"] * 0.3) + (filled_offices_df["primary_schools_1km"] *0.2) + (filled_offices_df["night_clubs_1km"] *0.3) + (filled_offices_df["starbucks_500m"] * (0.2))

filled_offices_df = filled_offices_df.sort_values("score", ascending=False)
filled_offices_df

top_3 = filled_offices_df.copy()

top_3 = top_3.head(3)

top_3

top_3.to_csv(r'output/top_3_offices.csv')

### Aquí y para poder pintar el mapa vamos a reutilizar un par de funciones (haciendo modificaciones, por eso las trabajamos aquí y no en el archivo .py en el que se encuentran) para obtener las coordenadas del aeropuerto, el Starbucks, el nightclub y la escuela más cercanos a la oficina elegida.

def nearest_place(coordinates, places, place_type):
    min_distance = None
    nearest_place = None
    for place in places:
        place_coordinates = place["geometry"]["location"]
        distance_in_meters = fl.meters_between_coordinates(coordinates[1], coordinates[0], place_coordinates["lat"], place_coordinates["lng"])
        if not min_distance:
            min_distance = distance_in_meters
            nearest_place = place
        elif distance_in_meters < min_distance:
            min_distance = distance_in_meters
            nearest_place = place
    
    if nearest_place:
        return {"name": nearest_place["name"], 
                "location": nearest_place["geometry"]["location"], 
                "meters_to_office": min_distance,
                "type": place_type}
    else:
        return {}

def fill_nearest_places(coordinates):
    latlong = str(coordinates[::-1]).replace(' ', '')[1:-1]
    nearest_places = []
    airports_nearby = gl.get_google(latlong, radius=2000, typ='airport', apiKey=apiKey, keyw='international')
    nearest_places.append(nearest_place(coordinates, airports_nearby, 'airport'))
    
    primary_schools_nearby = gl.get_google(latlong, radius=1000, typ='primary_school', apiKey=apiKey)
    nearest_places.append(nearest_place(coordinates, primary_schools_nearby, 'primary_school'))
    
    
    night_clubs_nearby = gl.get_google(latlong, radius=1000, typ='night_club', apiKey=apiKey)
    nearest_places.append(nearest_place(coordinates, night_clubs_nearby, 'night_club'))

    starbucks_nearby = gl.get_google(latlong, radius=500, typ='cafe', apiKey=apiKey, keyw='starbucks')
    nearest_places.append(nearest_place(coordinates, starbucks_nearby, 'starbucks'))

    return nearest_places

nearest_places = fill_nearest_places([-123.117650, 49.282159])

nearest_places
