==> 3.    Geoqueries.ipynb <==
from pymongo import MongoClient
import pandas as pd

import src.googla as gl
import src.geoqueries as g
import os
from dotenv import load_dotenv
import json
import requests
import src.filtering as fl
import src.geoqueries as ge
from pymongo import GEOSPHERE

load_dotenv()

### Para utilizar también MongoDB para hacer geoqueries, vamos a complementar el resultado obtenido en el ranking hecho anteriormente averiguando cuál de las tres oficinas con más puntuación está más cerca de una peluquería canina para Pepe, el perro de la oficina.

conn = MongoClient("localhost:27017")

db = conn.get_database("Cleaned_offices")

### Creamos el índice 2dsphere que soporta queries que calculan geometrías con la forma de la esfera terrestre. Así podemos calcular distancias y lanzar queries que nos permitan averiguar qué queda cerca de qué.

collection = db.get_collection("cleaned_offices")

collection.create_index([("location", '2dsphere')])




### Filtramos las oficinas de Canadá de nuevo para que los resultados de la geoquery sean coherentes con el análisis hecho hasta aquí.

can_offices = collection.find({"offices.country_code": 'CAN', "city": "Vancouver"})

can_offices

offices_list = list(can_offices)

db.collection.create_index([("location", '2dsphere')])

#collection.find_one()

#db.collection.index_information()

### Realizamos una llamada a la API de FourSquare para obtener los locales de Pet Services para obtener sus coordenadas y así averiguar si hay alguno cercano mediante la geoquery.

client_id = os.getenv("API_FS_CL")
secret = os.getenv("API_FS_KEY")

enlace = 'https://api.foursquare.com/v2/venues/explore'
pet_service = "5032897c91d4c4b30a586d69"



vancouver = {'type': 'Point', 'coordinates': [49.1902, -123.1837]}



parametros = { "client_id" : client_id,
              "client_secret" : secret,
              "v" : "20180323",
              "ll": f"{vancouver.get('coordinates')[0]},{vancouver.get('coordinates')[1]}",
              "categoryId": pet_service,
              "limit" : 100
}

resp = requests.get (url = enlace, params = parametros)
pet_service = json.loads(resp.text)

pet_service.keys()

### A partir de aquí y partiendo de las keys que ya hemos visto vamos a intentar obtener el nombre del local, la ciudad a la que pertenece (para asegurarnos de que es Vancouver así como la latitud y la longitud de cada uno).

res = pet_service.get('response')
#res



groups = res.get("groups")

items = groups[0].get("items")

from functools import reduce
import operator

def getFromDict(diccionario,mapa):
    return reduce(operator.getitem,mapa,diccionario)

nombre = ["venue", "name"]
ciudad = ["venue","location", "city"]
latitud = ["venue", "location","lat"]
longitud = ["venue","location","lng"]



obtained_data = []
for diccionario in items:
    empty = {}
    empty["nombre"] = getFromDict(diccionario,nombre)
    empty["ciudad"] = getFromDict(diccionario,ciudad)
    empty["latitud"] = getFromDict(diccionario,latitud)
    empty["longitud"] = getFromDict(diccionario,longitud)
    obtained_data.append(empty)

obtained_data

pet_services = pd.DataFrame(obtained_data)

pet_services

### Aplicamos la función utilizada anteriormente para obtener la localización tipo "Point".

pet_services["location"] = pet_services[["latitud","longitud"]].apply(lambda x:fl.toGeoJson(x.latitud,x.longitud), axis=1)
pet_services.head()

#pet_services.to_json("pet_services.json", orient="records")

### Pasamos el dataframe a Json y lo importamos en Mongo DB para poder trabajar con la geoquery.

#!mongoimport --db companies --collection pet_services --drop --jsonArray output/pet_services.json

conn = MongoClient("localhost:27017")

db1 = conn.get_database("companies")

collection = db1.get_collection("pet_services")

collection.create_index([("location", '2dsphere')])

collection.create_index([("geometry", '2dsphere')])

collection.find_one().keys()

#collection.find_one()

### Importamos el CSV con el dataframe formado por las 3 oficinas con mejor puntuación para obtener sus coordenadas y poder hacer la geoquery.

top3_offices = pd.read_csv('output/top_3_offices.csv')

top3_offices

### Aplicamos la función "find_pet_services" que realiza una geoquery con las coordenadas de las oficinas para saber si existe algún local de servicios para perros en un radio de 4 km.

top3_offices['pet_services'] = top3_offices.apply(lambda x: ge.find_pet_services(db1, [x['longitude'], x['latitude']]), axis=1)

top3_offices['pet_services'][0]

top3_offices['nearest_pet_service'] = top3_offices.apply(lambda x: gl.nearest_place(x['longitude'], x['latitude'], x['pet_services']), axis=1)

### Con esta función creamos la columna nearest_pet_service por cada una de las tres oficinas que nos muestra de la información de los locales de cuidado de mascotas, el más cercano para cada oficina.

top3_offices

### Descartamos la 3 oficina definitivamente porque no tiene cerca una peluquería canina así que vamos a analizar de entre las otras dos, cuál tiene una más cerca:

top3_offices['nearest_pet_service'][0]

top3_offices['nearest_pet_service'][1]

### Como podemos comprobar, la oficina con índice 0, la ganadora del ranking anterior es la que tiene más cerca el local de servicios para mascotas, así que sigue siendo la elegida. A continuación, en el siguiente .ipynb vamos a pintar un mapa con Folium para ver cómo quedaría el mapa con: la oficina elegida con todos los servicios así como la que ha quedado en segunda posición

## En conclusión,nos quedamos con la oficina 'Nexterra' de la categoría 'cleantech' y con las siguientes coordenadas: lat = 49.282159, lon =	-123.117650.
