==> 4.   Folium.ipynb <==
import folium
from folium import Choropleth, Circle, Marker, Icon, Map
from folium.plugins import HeatMap, MarkerCluster
import pandas as pd

### Ahora vamos paso por paso añadiendo los diferentes markers al mapa para mostrar las localizaciones. La estructura siempre es la misma:
 - Declaramos las variables latitud y longitud
 - declaramos el mapa al que le pasamos la localización, el icono y el "tooltip" o etiqueta que aparece cuando  pasamos el cursor por encima
    del icono.
    
#### El zoom que queremos darle al mapa al pintarse lo declararemos junto al primer mapa  al cual se van añadiendo los markers posteriormente.

inicial_lat = 49.282159
inicial_long = -123.117650

map_1 = folium.Map(location = [inicial_lat,inicial_long], zoom_start = 15)
map_1

iconito = Icon(color = "blue",
             prefix = "fa",
             icon = "building",
             icon_color = "black",
             tooltip = "Office 1"
)

#Marker Office 1
office1 = Marker(location=[inicial_lat,inicial_long], tooltip = "Oficina elegida", icon = iconito)
office1.add_to(map_1)
map_1

#Ubicación office2
office2_lat = 49.283935
office2_long = -123.115930

icono = Icon(color = "red",
             prefix = "fa",
             icon = "building",
             icon_color = "black",
)
loc = [office2_lat, office2_long]

marker_office2 = Marker(location = loc, icon = icono, tooltip = "Oficina segunda")

marker_office2.add_to(map_1)
map_1

#Ubicación del aeropuerto más cercano. A 639.51190 metros.

airport_lat = 49.2842096 
airport_long = -123.1094381

icono = Icon(color = "green",
             prefix = "fa",
             icon = "plane",
             icon_color = "black",
)
loc = [airport_lat, airport_long]


marker_airport = Marker(location = loc, icon = icono, tooltip = "YVR International Airport")
marker_airport.add_to(map_1)
map_1

#Ubicación del Starbucks más cercano. A 58.18427853798812 metros.

starbucks_lat = 49.28207450000001 
starbucks_long = -123.1184392

icono = Icon(color = "orange",
             prefix = "fa",
             icon = "coffee",
             icon_color = "black",
)
loc = [starbucks_lat, starbucks_long]



marker_starbucks = Marker(location = loc, icon = icono, tooltip = "Starbucks")
marker_starbucks.add_to(map_1)
map_1

#Ubicación del colegio más cercano. A 591.948925 metros.

school_lat = 49.277642 
school_long = -123.1133467

icono = Icon(color = "pink",
             prefix = "fa",
             icon = "pencil",
             icon_color = "black",
)
loc = [school_lat, school_long]




marker_school = Marker(location = loc, icon = icono, tooltip = "School")
marker_school.add_to(map_1)
map_1

#Ubicación del club más cercano. A 286.9911 metros.

club_lat = 49.2845532 
club_long = -123.1191216

icono = Icon(color = "black",
             prefix = "fa",
             icon = "glass",
             icon_color = "white",
)
loc = [club_lat, club_long]

marker_club = Marker(location = loc, icon = icono, tooltip = "Night Club")
marker_club.add_to(map_1)
map_1

#Ubicación del local de servicios para mascotas más cercano. A 3401.02217 metros.

pets_lat = 49.2522550396 
pets_long = -123.1274256764

icono = Icon(color = "white",
             prefix = "fa",
             icon = "paw",
             icon_color = "cadetblue",
)
loc = [pets_lat, pets_long]

# 🗺🗺🗺 AQUÍ TENEMOS EL MAPA 🗺🗺🗺

marker_pets = Marker(location = loc, icon = icono, tooltip = "Pet Services")
marker_pets.add_to(map_1)
map_1
